feat(Buttons): Criação das declarações de modulos e componentes
