+++ src/components/Buttons/Deprecated/Sh3SubmitButton.vue
@@ -4,7 +4,7 @@
     :class="
       twMerge(
         'h-[55px] w-full py-5 inline-flex justify-center items-center gap-2.5 rounded-lg  text-white bg-blue-800 hover:bg-blue-700 active:bg-blue-500',
-        submitButton.addStyle,
+        submitButton.addStyle
       )
     "
   >
@@ -15,7 +15,12 @@
 <script setup lang="ts">
 import type { PropType } from "vue";
 import { twMerge } from "tailwind-merge";
-import type { SubmitButton } from "../types";
+
+export type SubmitButton = {
+  form: string;
+  addStyle: string;
+  label: string;
+};
 
 defineProps({
   submitButton: {
